move rebranding to root pages
