+++ docs/pages/index.tsx
@@ -1,161 +1,43 @@
 import * as React from 'react';
-import { styled } from '@material-ui/core/styles';
-import Box from '@material-ui/core/Box';
-import Typography, { TypographyProps } from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
-import Container from '@material-ui/core/Container';
-import Steps from 'docs/src/pages/landing/Steps';
-import Themes from 'docs/src/pages/landing/Themes';
-import QuickWord from 'docs/src/pages/landing/QuickWord';
-import Sponsors from 'docs/src/pages/landing/Sponsors';
-import Users from 'docs/src/pages/landing/Users';
-import Quotes from 'docs/src/pages/landing/Quotes';
-import Pro from 'docs/src/pages/landing/Pro';
-import AppFooter from 'docs/src/modules/components/AppFooter';
-import AppFrame from 'docs/src/modules/components/AppFrame';
-import Link from 'docs/src/modules/components/Link';
 import Head from 'docs/src/modules/components/Head';
-import loadScript from 'docs/src/modules/utils/loadScript';
-import { useTranslate } from 'docs/src/modules/utils/i18n';
+import NoSsr from '@material-ui/core/NoSsr';
+import Divider from '@material-ui/core/Divider';
+import AppHeader from 'docs/src/layouts/AppHeader';
+import Hero from 'docs/src/components/home/Hero';
+import References, { CORE_CUSTOMERS } from 'docs/src/components/home/References';
+import ProductSuite from 'docs/src/components/home/ProductSuite';
+import ValueProposition from 'docs/src/components/home/ValueProposition';
+import DesignSystemComponents from 'docs/src/components/home/DesignSystemComponents';
+import Testimonials from 'docs/src/components/home/Testimonials';
+import Sponsors from 'docs/src/components/home/Sponsors';
+import HeroEnd from 'docs/src/components/home/HeroEnd';
+import AppFooter from 'docs/src/layouts/AppFooter';
+import BrandingProvider from 'docs/src/BrandingProvider';
+import NewsletterToast from 'docs/src/components/home/NewsletterToast';
 
-let dependenciesLoaded = false;
-
-function loadDependencies() {
-  if (dependenciesLoaded) {
-    return;
-  }
-
-  dependenciesLoaded = true;
-
-  loadScript('https://buttons.github.io/buttons.js', document.querySelector('head'));
-  loadScript('https://platform.twitter.com/widgets.js', document.querySelector('head'));
-}
-
-const Content = styled(Container)(({ theme }) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  textAlign: 'center',
-  paddingTop: theme.spacing(4),
-  paddingBottom: theme.spacing(8),
-  [theme.breakpoints.up('md')]: {
-    paddingTop: theme.spacing(16),
-    paddingBottom: theme.spacing(16),
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    textAlign: 'left',
-  },
-}));
-
-const Title = styled(Typography)<TypographyProps & { component?: React.ElementType }>(
-  ({ theme }) => ({
-    marginLeft: -12,
-    whiteSpace: 'nowrap',
-    letterSpacing: '.7rem',
-    textIndent: '.7rem',
-    fontWeight: theme.typography.fontWeightLight,
-    [theme.breakpoints.only('xs')]: {
-      fontSize: 28,
-    },
-  }),
-);
-
-const Logo = styled('img')(({ theme }) => ({
-  flexShrink: 0,
-  width: 120,
-  height: 120,
-  marginBottom: theme.spacing(2),
-  [theme.breakpoints.up('md')]: {
-    marginRight: theme.spacing(8),
-    width: 195,
-    height: 175,
-  },
-}));
-
-const Social = styled('div')(({ theme }) => ({
-  padding: theme.spacing(2, 0),
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  height: 20,
-  boxSizing: 'content-box',
-  '& a': {
-    color: theme.palette.background.paper,
-  },
-}));
-
-export default function LandingPage() {
-  React.useEffect(() => {
-    loadDependencies();
-  }, []);
-  const t = useTranslate();
-
+export default function Home() {
   return (
-    <AppFrame>
-      <Box sx={{ flex: '1 0 100%' }}>
-        <Head />
-        <main id="main-content" tabIndex={-1}>
-          <Box sx={{ pt: 8, color: 'primary.main' }}>
-            <Content maxWidth="md">
-              <Logo src="/static/logo_raw.svg" alt="" />
-              <div>
-                <Title variant="h3" component="h1" color="inherit" gutterBottom>
-                  {'MATERIAL-UI'}
-                </Title>
-                <Typography variant="h5" component="p" color="inherit">
-                  {t('strapline')}
-                </Typography>
-                <Button
-                  component={Link}
-                  noLinkStyle
-                  href="/getting-started/installation"
-                  sx={{ mt: 4 }}
-                  variant="outlined"
-                >
-                  {t('getStarted')}
-                </Button>
-              </div>
-            </Content>
-          </Box>
-          <Social>
-            <Box
-              sx={{
-                width: 105,
-                display: 'flex',
-                justifyContent: 'flex-end',
-                mr: 1,
-                '& span': { display: 'flex' },
-              }}
-            >
-              <a
-                className="github-button"
-                href="https://github.com/mui-org/material-ui"
-                data-icon="octicon-star"
-                data-show-count="true"
-              >
-                Star
-              </a>
-            </Box>
-            <Box sx={{ width: 160, display: 'flex' }}>
-              <a
-                className="twitter-follow-button"
-                href="https://twitter.com/@materialui"
-                data-show-screen-name="false"
-              >
-                Follow
-              </a>
-            </Box>
-          </Social>
-          <Pro />
-          <QuickWord />
-          <Steps />
-          <Themes />
-          <Sponsors />
-          <Quotes />
-          <Users />
-        </main>
-        <AppFooter />
-      </Box>
+    <BrandingProvider>
+      <Head
+        title="MUI: A popular React UI framework"
+        description="The ultimate solution for your UI. MUI provides a robust, customizible and accessible library of foundational and advanced components, enabling you to build your own design system and develop React applications faster."
+      />
+      <NoSsr>
+        <NewsletterToast />
+      </NoSsr>
+      <AppHeader />
+      <main>
+        <Hero />
+        <References companies={CORE_CUSTOMERS} />
+        <ProductSuite />
+        <ValueProposition />
+        <DesignSystemComponents />
+        <Testimonials />
+        <Sponsors />
+        <HeroEnd />
+        <Divider />
+      </main>
+      <AppFooter />
       <script
         type="application/ld+json"
         // eslint-disable-next-line react/no-danger
@@ -176,6 +58,6 @@ export default function LandingPage() {
           `,
         }}
       />
-    </AppFrame>
+    </BrandingProvider>
   );
 }
