+++ docs/pages/pricing.tsx
