+++ docs/src/route.ts
@@ -1,5 +1,5 @@
 const ROUTES = {
-  home: '/branding/home',
+  home: '/',
   productCore: '/branding/core/',
   productAdvanced: '/branding/x-advanced/',
   productTemplates: '/branding/templates/',
@@ -10,12 +10,12 @@ const ROUTES = {
   customization: '/customization/how-to-customize/',
   theming: '/customization/theming/',
   documentation: '/getting-started/usage/',
-  pricing: '/branding/pricing/',
+  pricing: '/pricing/',
   blog: '/blog/',
   showcase: '/discover-more/showcase',
   roadmap: '/discover-more/roadmap',
   languages: '/discover-more/languages',
-  about: '/branding/about/',
+  about: '/about/',
   vision: '/discover-more/vision/',
   careers: '/company/careers/',
   support: '/getting-started/support/#professional-support-premium',
